+++ web/download_manager.js
@@ -14,7 +14,7 @@
  */
 
 import { createObjectURL, createValidAbsoluteUrl, isPdfFile } from "pdfjs-lib";
-import { viewerCompatibilityParams } from "./viewer_compatibility.js";
+import { compatibilityParams } from "./app_options.js";
 
 if (typeof PDFJSDev !== "undefined" && !PDFJSDev.test("CHROME || GENERIC")) {
   throw new Error(
@@ -58,7 +58,7 @@ class DownloadManager {
     const blobUrl = createObjectURL(
       data,
       contentType,
-      viewerCompatibilityParams.disableCreateObjectURL
+      compatibilityParams.disableCreateObjectURL
     );
     download(blobUrl, filename);
   }
@@ -70,7 +70,7 @@ class DownloadManager {
     const isPdfData = isPdfFile(filename);
     const contentType = isPdfData ? "application/pdf" : "";
 
-    if (isPdfData && !viewerCompatibilityParams.disableCreateObjectURL) {
+    if (isPdfData && !compatibilityParams.disableCreateObjectURL) {
       let blobUrl = this._openBlobUrls.get(element);
       if (!blobUrl) {
         blobUrl = URL.createObjectURL(new Blob([data], { type: contentType }));
@@ -113,7 +113,7 @@ class DownloadManager {
    *   the "open with" dialog.
    */
   download(blob, url, filename, sourceEventType = "download") {
-    if (viewerCompatibilityParams.disableCreateObjectURL) {
+    if (compatibilityParams.disableCreateObjectURL) {
       // URL.createObjectURL is not supported
       this.downloadUrl(url, filename);
       return;
